refactor: restricting access to user slip data

+++ src/repositories/BoletoRepository.ts
@@ -6,31 +6,38 @@ interface BoletoType {
    vencimento: Date;
    valor: string;
    codigo: string;
+   user: string;
 }
 
 class BoletoRepository {
-   async findOne(_id: string) {
-      return await Boleto.findById(_id);
+   async findOne(_id: string, userId: string) {
+      return await Boleto.findById(_id).where("user").equals(userId);
    }
 
-   async findAll() {
-      return await Boleto.find();
+   async findAll(userId: string) {
+      return await Boleto.find().where("user").equals(userId);
    }
 
-   async findAllPaid() {
-      return await Boleto.find().where("status").equals(StatusBoleto["PAGO"]);
+   async findAllPaid(userId: string) {
+      return await Boleto.find()
+         .where("status")
+         .equals(StatusBoleto["PAGO"])
+         .where("user")
+         .equals(userId);
    }
 
    async create(boleto: BoletoType) {
       await Boleto.create(boleto);
    }
 
-   async update(_id: string) {
-      await Boleto.findByIdAndUpdate(_id, { status: StatusBoleto.PAGO });
+   async update(_id: string, userId: string) {
+      await Boleto.findByIdAndUpdate(_id, { status: StatusBoleto.PAGO })
+         .where("user")
+         .equals(userId);
    }
 
-   async deleteOne(_id: string) {
-      await Boleto.findByIdAndDelete(_id);
+   async deleteOne(_id: string, userId: string) {
+      await Boleto.findByIdAndDelete(_id).where("user").equals(userId);
    }
 }
